Adding descriptive documentation
