+++ Sources/ForageSDK/Component/ForagePANTextField/Protocol/ForagePANTextFieldDelegate.swift
@@ -7,15 +7,20 @@
 
 import UIKit
 
-// MARK: - ForagePANTextFieldDelegate
-
-/// Protocol to comunicate `ForagePANTextField` to the client application
-public protocol ForagePANTextFieldDelegate: AnyObject {
-    func panNumberStatus(_ view: UIView, cardStatus: CardStatus)
-}
-
 public enum CardStatus: String {
     case valid
     case invalid
     case identifying
+}
+
+// MARK: - ForagePANTextFieldDelegate
+
+/// Protocol to comunicate `ForagePANTextField` to the client application
+public protocol ForagePANTextFieldDelegate: AnyObject {
+    /// Ebt card number status
+    ///
+    /// - Parameters:
+    ///  - view: ForagePANTextField view reference.
+    ///  - cardStatus: Ebt card number validations status. Check ``CardStatus`` for details.
+    func panNumberStatus(_ view: UIView, cardStatus: CardStatus)
 }
